#672 Private classes moved to their own files. File renamed to FileRanges because File is too common.
