+++ Assets/SEE/DataModel/DG/SourceRange/FileRanges.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 34f1a3c93ea821546a8621715396b14a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
