+++ Assets/SEE/DataModel/DG/SourceRange/Range.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 103d74aad39844c469200b158df07b66
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
